optimized code, refactored by implementing dto, service and controller, object mapper for response
