+++ src/main/resources/application.properties
@@ -1,1 +1,1 @@
-
+spring.jackson.serialization.indent_output=true
