+++ src/test/java/com/codechefcards/api/CodeChefCardsApiApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class CodechefCardsApiApplicationTests {
+class CodeChefCardsApiApplicationTests {
 
 	@Test
 	void contextLoads() {
